requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.2
django-jazzmin==3.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
environs==14.6.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
marshmallow==4.2.2
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
